requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
filelock==3.18.0
Flask==3.1.0
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
mysqlclient==2.2.7
platformdirs==4.3.7
PyMySQL==1.1.1
virtualenv==20.30.0
Werkzeug==3.1.3
WTForms==3.2.1
